reorganize google auth & sync code
